Cambios en los DTOs: los campos tipo String id fueron cambiados a Long

+++ src/main/java/com/avanzada/alojamientos/DTO/CommentDTO.java
@@ -3,14 +3,14 @@ package com.avanzada.alojamientos.DTO;
 import java.time.LocalDateTime;
 
 public record   CommentDTO (
-        Integer id,
+        long id,
         Integer rating,
         String text,
         LocalDateTime createdAt,
         Boolean isModerated,
         Integer reservationId,
         Integer accommodationId,
-        Integer userId,
+        long userId,
         ReplyHostDTO hostReply
 ){
